CoFi-Dec-main/condaenv.4k16no3e.requirements.txt
absl-py==2.1.0
accelerate==0.27.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
av==14.1.0
backoff==2.2.1
beautifulsoup4==4.13.3
bitsandbytes==0.41.0
black==25.1.0
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
boltons==21.0.0
braceexpand==0.1.7
bracex==2.5.post1
brotli==1.1.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-option-group==0.5.6
cloudpathlib==0.20.0
cmake==3.31.4
codeshield==1.0.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contexttimer==0.3.3
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.3.2
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
deprecated==1.2.18
diffusers==0.21.0
dill==0.3.8
distlib==0.3.9
distro==1.9.0
effdet==0.4.1
einops==0.8.1
einops-exts==0.0.4
emoji==2.14.1
eval-type-backport==0.2.2
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.2.0
face==24.0.0
fairscale==0.4.13
faiss-gpu==1.7.2
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
filetype==1.2.0
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.12
gitpython==3.1.44
glom==22.1.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-vision==3.10.0
googleapis-common-protos==1.68.0
gradio==3.35.2
gradio-client==0.2.9
grpcio==1.70.0
grpcio-status==1.63.0rc1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
humanfriendly==10.0
identify==2.6.7
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==7.1.0
inflate64==1.0.1
iopath==0.1.10
ipython==8.32.0
jedi==0.19.2
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.6.17
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language-data==1.3.0
lazy-loader==0.4
linkify-it-py==2.0.3
lit==18.1.8
llama-recipes==0.0.5.post2
llava==0.0.1.dev0
loralib==0.1.2
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==2.2.0
markupsafe==2.0.1
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
modelscope==1.23.1
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multivolumefile==0.2.3
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.27.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
open-clip-torch==2.31.0
openai==1.64.0
opencv-python==4.11.0.86
opendatasets==0.1.22
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
optimum==1.24.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pdfminer-six==20240706
peewee==3.17.9
peft==0.6.0
pexpect==4.9.0
pi-heif==0.21.0
pikepdf==9.5.2
pillow==10.4.0
platformdirs==4.3.6
plotly==6.0.0
portalocker==3.1.1
pre-commit==4.1.0
preshed==3.0.9
proglog==0.1.10
prompt-toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.0
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py7zr==0.22.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pybcj==1.0.3
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic-core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pygments==2.19.1
pyparsing==3.2.1
pypdf==5.3.0
pypdfium2==4.30.1
pyppmd==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-slugify==8.0.4
pytz==2025.1
pyyaml==6.0.1
pyzstd==0.16.2
rapidfuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.5.3
rouge-score==0.1.2
rpds-py==0.23.1
rsa==4.9
ruamel-yaml==0.18.10
ruamel-yaml-clib==0.2.12
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.2.2
scipy==1.15.2
semantic-version==2.10.0
semgrep==1.109.0
sentence-transformers==3.4.1
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.3
streamlit==1.42.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
texttable==1.7.0
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.9.0
timm==1.0.14
tokenize-rt==6.1.0
tokenizers==0.13.3
toml==0.10.2
torch==2.6.0
torchaudio==2.1.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typer==0.15.1
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
uc-micro-py==1.0.3
unstructured==0.16.23
unstructured-client==0.30.3
unstructured-inference==0.8.7
unstructured-pytesseract==0.3.13
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
wasabi==1.1.3
watchdog==6.0.0
wcmatch==8.5.2
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.111
webencodings==0.5.1
websockets==15.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0